+++ src/injectionSymbols.ts
@@ -1,4 +1,4 @@
-import { InjectionKey, ComputedRef } from 'vue'
+import { InjectionKey, ComputedRef, Ref } from 'vue'
 import { RouteLocationNormalizedLoaded } from './types'
 import { Router } from './router'
 import { RouteRecordNormalized } from './matcher/types'
@@ -13,16 +13,30 @@ export const PolySymbol = (name: string) =>
     : (__DEV__ ? '[vue-router]: ' : '_vr_') + name
 
 // rvlm = Router View Location Matched
+/**
+ * RouteRecord being rendered by the closest ancestor Router View. Used for
+ * `onBeforeRouteUpdate` and `onBeforeRouteLeave`. rvlm stands for Router View
+ * Location Matched
+ *
+ * @internal
+ */
 export const matchedRouteKey = /*#__PURE__*/ PolySymbol(
   __DEV__ ? 'router view location matched' : 'rvlm'
 ) as InjectionKey<ComputedRef<RouteRecordNormalized | undefined>>
-// rvd = Router View Depth
+
+/**
+ * Allows overriding the router view depth to control which component in
+ * `matched` is rendered. rvd stands for Router View Depth
+ *
+ * @internal
+ */
 export const viewDepthKey = /*#__PURE__*/ PolySymbol(
   __DEV__ ? 'router view depth' : 'rvd'
 ) as InjectionKey<number>
 
 /**
- * Allows overriding the router instance returned by `useRouter` in tests. r stands for router
+ * Allows overriding the router instance returned by `useRouter` in tests. r
+ * stands for router
  *
  * @internal
  */
@@ -31,10 +45,21 @@ export const routerKey = /*#__PURE__*/ PolySymbol(
 ) as InjectionKey<Router>
 
 /**
- * Allows overriding the current route returned by `useRoute` in tests. rl stands for route location
+ * Allows overriding the current route returned by `useRoute` in tests. rl
+ * stands for route location
  *
  * @internal
  */
 export const routeLocationKey = /*#__PURE__*/ PolySymbol(
   __DEV__ ? 'route location' : 'rl'
 ) as InjectionKey<RouteLocationNormalizedLoaded>
+
+/**
+ * Allows overriding the current route used by router-view. Internally this is
+ * used when the `route` prop is passed.
+ *
+ * @internal
+ */
+export const routerViewLocationKey = /*#__PURE__*/ PolySymbol(
+  __DEV__ ? 'router view location' : 'rvl'
+) as InjectionKey<Ref<RouteLocationNormalizedLoaded>>